feat: modify url path and http method
